+++ android/app/src/main/java/com/h19h29/naymnaymlevelup/rebuild/migration/RebuildMigrationCoordinator.kt
@@ -72,7 +72,7 @@ data class MigrationVerification(
     val profileCount: Int,
     val mealRecordCount: Int,
     val progressEventCount: Int,
-    val totalXp: Int,
+    val totalXp: Long,
 )
 
 interface MigrationTarget {
